Moved default settings to folder
