+++ rendertron/middleware/django.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.http import HttpResponse
 
-from rendertron import default_settings
+from rendertron.settings import default_settings
 from rendertron.middleware.base import RendertronMiddleware
